Pull TailLogGroup Wizard into its own class
